ai-smartconnect4u-jitsi/requirements.txt
azure-functions
azure-data-tables
azure-storage-blob
azure-storage-queue
openai>=1.12.0
requests
python-dotenv
